move QuickCacheHelper to own module
